+++ test/exec/rawdofmt.c
@@ -20,7 +20,7 @@ int main(void)
 
     printf("Checking NewRawDoFmt...\n");
     NewRawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu", (VOID_FUNC)RAWFMTFUNC_COUNT, &count, "one", "two", 3, 4, 5, 6);
-    printf("Count is %u\n", count);
+    printf("Count is %u\n", (unsigned)count);
     NewRawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu", (VOID_FUNC)RAWFMTFUNC_STRING, buf, "one", "two", 3, 4, 5, 6);
     printf("Formatted string is: %s\n", buf);
     NewRawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu\n", (VOID_FUNC)RAWFMTFUNC_SERIAL, NULL, "one", "two", 3, 4, 5, 6);
@@ -29,7 +29,7 @@ int main(void)
     count = 0;
     printf("Checking RawDoFmt...\n");
     RawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu", args, (VOID_FUNC)RAWFMTFUNC_COUNT, &count);
-    printf("Count is %u\n", count);
+    printf("Count is %u\n", (unsigned)count);
     RawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu", args, (VOID_FUNC)RAWFMTFUNC_STRING, buf);
     printf("Formatted string is: %s\n", buf);
     RawDoFmt("%s plus %s will be %lu, next are %lu, %lu, %lu\n", args, (VOID_FUNC)RAWFMTFUNC_SERIAL, NULL);
